+++ src/cookie_agent/reward/__init__.py
@@ -1,0 +1,11 @@
+"""Public Reward Engine package exports."""
+
+from cookie_agent.reward.engine import RewardEngine
+from cookie_agent.reward.exceptions import RewardError
+from cookie_agent.reward.rule import RewardRule
+
+__all__: list[str] = [
+    "RewardEngine",
+    "RewardError",
+    "RewardRule",
+]